requirements.txt
dnspython>=2.3.0
requests>=2.28.0
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
responses>=0.23.0
